requirements.txt
Django==2.2.1
django-crispy-forms==1.7.2
djangorestframework==3.9.4
names==0.3.0
Pillow==6.0.0
pytz==2019.1
sqlparse==0.3.0
